+++ android/src/main/kotlin/com/oguzhnatly/flutter_android_auto/FAAEnums.kt
@@ -17,4 +17,5 @@ enum class FAAChannelTypes {
     onListItemSelectedComplete,
     onScreenBackButtonPressed,
     updateMessageTemplate,
+    updateLongMessageTemplate,
 }