+++ lib/aa_models/message/all.dart
@@ -1,1 +1,2 @@
+export 'long_message_template.dart';
 export 'message_template.dart';