+++ lib/android_auto_worker.dart
@@ -38,38 +38,37 @@ class FlutterAndroidAuto {
     _eventBroadcast = _androidAutoController.eventChannel
         .receiveBroadcastStream()
         .listen((event) {
-          final FAAChannelTypes receivedChannelType = EnumUtils.enumFromString(
-            FAAChannelTypes.values,
-            event['type'],
-          );
+      final FAAChannelTypes receivedChannelType = EnumUtils.enumFromString(
+        FAAChannelTypes.values,
+        event['type'],
+      );
 
-          switch (receivedChannelType) {
-            case FAAChannelTypes.onAndroidAutoConnectionChange:
-              final ConnectionStatusTypes connectionStatus =
-                  EnumUtils.enumFromString(
-                    ConnectionStatusTypes.values,
-                    event['data']['status'],
-                  );
-              _connectionStatus = connectionStatus.name;
-              if (_onAndroidAutoConnectionChange != null) {
-                _onAndroidAutoConnectionChange!(connectionStatus);
-              }
-              break;
-            case FAAChannelTypes.onListItemSelected:
-              _androidAutoController.processFAAListItemSelectedChannel(
-                event['data']['elementId'],
-              );
-              break;
-            case FAAChannelTypes.onScreenBackButtonPressed:
-              FlutterAndroidAutoController.templateHistory.removeWhere(
-                (AATemplate item) =>
-                    item.uniqueId == event['data']['elementId'],
-              );
-              break;
-            default:
-              break;
+      switch (receivedChannelType) {
+        case FAAChannelTypes.onAndroidAutoConnectionChange:
+          final ConnectionStatusTypes connectionStatus =
+              EnumUtils.enumFromString(
+            ConnectionStatusTypes.values,
+            event['data']['status'],
+          );
+          _connectionStatus = connectionStatus.name;
+          if (_onAndroidAutoConnectionChange != null) {
+            _onAndroidAutoConnectionChange!(connectionStatus);
           }
-        });
+          break;
+        case FAAChannelTypes.onListItemSelected:
+          _androidAutoController.processFAAListItemSelectedChannel(
+            event['data']['elementId'],
+          );
+          break;
+        case FAAChannelTypes.onScreenBackButtonPressed:
+          FlutterAndroidAutoController.templateHistory.removeWhere(
+            (AATemplate item) => item.uniqueId == event['data']['elementId'],
+          );
+          break;
+        default:
+          break;
+      }
+    });
   }
 
   /// A function that will disconnect all event listeners from Android Auto. The action
@@ -116,9 +115,9 @@ class FlutterAndroidAuto {
   static Future<void> setRootTemplate({required AATemplate template}) async {
     final bool? isCompleted = await _androidAutoController
         .flutterToNativeModule(FAAChannelTypes.setRootTemplate, {
-          'template': template.toJson(),
-          'runtimeType': _getAARuntimeTypeString(template),
-        });
+      'template': template.toJson(),
+      'runtimeType': _getAARuntimeTypeString(template),
+    });
 
     if (isCompleted == true) {
       if (FlutterAndroidAutoController.templateHistory.isEmpty) {
@@ -221,16 +220,16 @@ class FlutterAndroidAuto {
   /// Adds a template to the navigation hierarchy and displays it.
   ///
   /// - template is to add to the navigation hierarchy. **Must be one of the type:**
-  /// [AAGridTemplate] or [AAListTemplate] [AAInformationTemplat] [AAPointOfInterestTemplate] [AAMessageTemplate]
+  /// [AAGridTemplate] or [AAListTemplate] [AAInformationTemplat] [AAPointOfInterestTemplate] [AAMessageTemplate] [AALongMessageTemplate]
   ///
   /// Be aware of the restrictions : https://developer.android.com/training/cars/apps#template-restrictions
   /// - Max 5 templates in the navigation stack.
   static Future<bool> push({required AATemplate template}) async {
     final bool? isCompleted = await _androidAutoController
         .flutterToNativeModule(FAAChannelTypes.pushTemplate, <String, dynamic>{
-          'template': template.toJson(),
-          'runtimeType': _getAARuntimeTypeString(template),
-        });
+      'template': template.toJson(),
+      'runtimeType': _getAARuntimeTypeString(template),
+    });
     if (isCompleted == true) {
       FlutterAndroidAutoController.templateHistory.add(template);
     }
@@ -246,6 +245,7 @@ class FlutterAndroidAuto {
   static String _getAARuntimeTypeString(AATemplate template) {
     if (template is AAListTemplate) return 'FAAListTemplate';
     if (template is AAMessageTemplate) return 'FAAMessageTemplate';
+    if (template is AALongMessageTemplate) return 'FAALongMessageTemplate';
     return 'FAA${template.runtimeType}';
   }
 }