+++ lib/constants/private_constants.dart
@@ -44,4 +44,5 @@ enum FAAChannelTypes {
   onListItemSelectedComplete,
   onScreenBackButtonPressed,
   updateMessageTemplate,
+  updateLongMessageTemplate,
 }